Extract common PR/group parsing logic.
pr:deriveGroups
